move lower_cinn_fusion_op_pass
